Features/feature_2: Renamed folder with CamelCase and added new components  - added HeaderNotification component;  - added test skeletons for Header and HeaderNotification components;  - renamed components folders.  - refactored styles - sorted styles alphabetically.
