+++ src/components/Header/Header.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Header from './Header';
+
+it('renders Header without crashing', () => {
+	const div = document.createElement('div');
+	ReactDOM.render(<Header />, div);
+	ReactDOM.unmountComponentAtNode(div);
+});
