+++ src/styles/components/_header.scss
@@ -1,68 +1,74 @@
 $header-background-color: #393535;
 $header-title-color: #fff;
-$notification-color: red;
+$notification-color: #f00;
 $header-notification-color: #fff;
 
 .header {
-	background-color: $header-background-color;
+  background-color: $header-background-color;
 
-	display: flex;
-	flex-direction: column;
+  display: flex;
+  flex-direction: column;
 
-	position: fixed;
-	top: 0;
-	left: 0;
-	height: 3rem;
-	width: 100vw;
+  height: 3rem;
+  left: 0;
+  opacity: .8;
+  position: fixed;
+  text-align: center;
+  top: 0;
+  width: 100vw;
 
-	text-align: center;
-	opacity: 0.8;
+  .header-title {
+    color: $header-title-color;
+    font: 'Garamond';
+    font-size: 2rem;
+    line-height: 3rem;
+  }
 
-	.header-title {
-		color: $header-title-color;
-		font: 'Garamond';
-		font-size: 2rem;
-		line-height: 3rem;
-	}
+  .header-action {
+    background: none;
+    border: none;
+    bottom: 0rem;
 
-	.header-action {
-		font-size: 1rem;
-		background: none;
-		color: $header-title-color;
+    color: $header-title-color;
 
-		position: absolute;
-		top: 0rem;
-		bottom: 0rem;
+    font-size: 1rem;
 
-		outline: none;
-		border: none;
-	}
+    outline: none;
 
-	.header-notification {
-		background: url('../media/notification.png');
-		background-size: contain;
+    position: absolute;
 
-		position: absolute;
-		top: 0.75rem;
-		right: 0.5rem;
+    top: 0rem;
+  }
+}
 
-		height: 1.75rem;
-		width: 1.75rem;
+.header-notification {
+  background: url('../media/notification.png');
+  background-size: contain;
 
-		.notification-counter {
-			font-size: 0.7rem;
-			font-weight: 300;
+  height: 1.75rem;
+  position: absolute;
 
-			background-color: $notification-color;
-			color: $header-notification-color;
+  right: .5rem;
+  top: .75rem;
 
-			border-radius: 2px;
-			height: 1rem;
-			width: 1rem; 
+  width: 1.75rem;
 
-			position: absolute;
-			bottom: -0.2rem;
-			right: -0.2rem;
-		}
-	}
+  .notification-counter {
+    background-color: $notification-color;
+    border-radius: 2px;
+    bottom: -.2rem;
+
+    color: $header-notification-color;
+
+    font-size: .7rem;
+    font-weight: 300;
+
+    height: 1rem;
+
+    min-width: 1rem;
+
+    position: absolute;
+
+    right: -.2rem;
+  }
 }
