+++ src/styles/index.css
@@ -39,46 +39,47 @@
   background-color: #393535;
   display: flex;
   flex-direction: column;
-  position: fixed;
-  top: 0;
-  left: 0;
   height: 3rem;
-  width: 100vw;
+  left: 0;
+  opacity: .8;
+  position: fixed;
   text-align: center;
-  opacity: 0.8; }
+  top: 0;
+  width: 100vw; }
   .header .header-title {
     color: #fff;
     font: 'Garamond';
     font-size: 2rem;
     line-height: 3rem; }
   .header .header-action {
-    font-size: 1rem;
     background: none;
-    color: #fff;
-    position: absolute;
-    top: 0rem;
+    border: none;
     bottom: 0rem;
+    color: #fff;
+    font-size: 1rem;
     outline: none;
-    border: none; }
-  .header .header-notification {
-    background: url("../media/notification.png");
-    background-size: contain;
     position: absolute;
-    top: 0.75rem;
-    right: 0.5rem;
-    height: 1.75rem;
-    width: 1.75rem; }
-    .header .header-notification .notification-counter {
-      font-size: 0.7rem;
-      font-weight: 300;
-      background-color: red;
-      color: #fff;
-      border-radius: 2px;
-      height: 1rem;
-      width: 1rem;
-      position: absolute;
-      bottom: -0.2rem;
-      right: -0.2rem; }
+    top: 0rem; }
+
+.header-notification {
+  background: url("../media/notification.png");
+  background-size: contain;
+  height: 1.75rem;
+  position: absolute;
+  right: .5rem;
+  top: .75rem;
+  width: 1.75rem; }
+  .header-notification .notification-counter {
+    background-color: #f00;
+    border-radius: 2px;
+    bottom: -.2rem;
+    color: #fff;
+    font-size: .7rem;
+    font-weight: 300;
+    height: 1rem;
+    min-width: 1rem;
+    position: absolute;
+    right: -.2rem; }
 
 .menu {
   padding-top: 15vh;
